LongDuZbot/requirements.txt
discord.py
requests
python-dotenv
Pillow
termcolor
pyfiglet
gitpython
schedule